# api/openapi.yaml
openapi: 3.1.0
info:
  title: Skawld Maintenance API
  version: 0.6.0-phase5
  description: >
    Phase 5 pilot-ready evidence-first maintenance intelligence contract.
    Learned candidates require exact evidence, human review, deterministic
    evaluation, approval authority, and explicit publication. AI quality
    labels are append-oriented. This API exposes no industrial control action
    or permit authority.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/workflow-candidates:
    post:
      summary: Compile a review-only workflow candidate
      description: >
        Uses at least two completed and approved semantic demonstrations.
        Compiler validation rejects fabricated event references, unknown tools,
        incomplete safety metadata, and unclassified semantic actions. This
        operation never publishes a workflow.
      tags: [Workflows]
      operationId: compileWorkflowCandidate
      security: [{sessionCookie: []}, {bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, demonstration_ids]
              properties:
                name: {type: string, minLength: 1, maxLength: 200}
                description: {type: string, maxLength: 2000}
                demonstration_ids:
                  type: array
                  minItems: 2
                  maxItems: 20
                  uniqueItems: true
                  items: {type: string, format: uuid}
      responses:
        "201":
          description: Review-only candidate compiled.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WorkflowVersion"}
        "400": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /api/v1/workflows:
    get:
      summary: List authorized workflow versions
      tags: [Workflows]
      operationId: listWorkflowVersions
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {in: query, name: cursor, required: false, schema: {type: string}}
        - {in: query, name: page_size, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
      responses:
        "200":
          description: Workflow versions in authorized site scope.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, next_cursor, has_more]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/WorkflowVersion"}
                  next_cursor: {type: [string, "null"]}
                  has_more: {type: boolean}
        "403": {$ref: "#/components/responses/Problem"}
  /api/v1/workflows/applicable:
    get:
      summary: List current published workflows applicable to one asset
      description: >
        Tenant, site, lifecycle, effective/review dates, and typed applicability
        are eligibility predicates in SQL before a workflow is returned.
      tags: [Workflows]
      operationId: listApplicableWorkflows
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - in: query
          name: asset_id
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Current applicable published workflow versions.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/WorkflowVersion"}
        "400": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
  /api/v1/workflows/{workflowID}/versions/{version}:
    get:
      summary: Get an exact workflow version and governance evidence
      tags: [Workflows]
      operationId: getWorkflowVersion
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/WorkflowID"}
        - {$ref: "#/components/parameters/WorkflowVersion"}
      responses:
        "200":
          description: Exact immutable candidate or published version.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WorkflowVersion"}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
  /api/v1/workflows/{workflowID}/versions/{version}/reviews:
    post:
      summary: Review an exact workflow candidate digest
      description: >
        Workflow review permission and a current scoped ApprovalAuthority are
        both required. Approval records applicability, prerequisites,
        competencies, effective date, and review date.
      tags: [Workflows]
      operationId: reviewWorkflowCandidate
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/WorkflowID"}
        - {$ref: "#/components/parameters/WorkflowVersion"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [decision, reason, applicability]
              properties:
                decision:
                  type: string
                  enum: [APPROVED, REJECTED, REVIEW_REQUIRED]
                reason: {type: string, minLength: 1, maxLength: 2000}
                applicability:
                  type: array
                  items: {$ref: "#/components/schemas/WorkflowApplicability"}
                prerequisites:
                  type: array
                  maxItems: 20
                  items: {type: string, minLength: 1, maxLength: 200}
                required_competencies:
                  type: array
                  maxItems: 20
                  items: {type: string, minLength: 1, maxLength: 200}
                effective_at: {type: string, format: date-time}
                review_at: {type: string, format: date-time}
      responses:
        "201":
          description: Append-only candidate review recorded.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WorkflowVersion"}
        "400": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /api/v1/workflows/{workflowID}/versions/{version}/publish:
    post:
      summary: Evaluate and publish an approved workflow candidate
      description: >
        Publication requires workflow-publish permission, a separate current
        ApprovalAuthority, an approved exact digest, unchanged tool contracts,
        and passing deterministic task-success and unsafe-action gates.
      tags: [Workflows]
      operationId: publishWorkflowVersion
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/WorkflowID"}
        - {$ref: "#/components/parameters/WorkflowVersion"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: {type: string, minLength: 1, maxLength: 2000}
      responses:
        "200":
          description: Immutable effective-dated workflow version published.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WorkflowVersion"}
        "403": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/workflows/{workflowID}/versions/{version}/applicability:
    post:
      summary: Approve a separate applicability expansion
      tags: [Workflows]
      operationId: expandWorkflowApplicability
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/WorkflowID"}
        - {$ref: "#/components/parameters/WorkflowVersion"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason, applicability]
              properties:
                reason: {type: string, minLength: 1, maxLength: 2000}
                applicability:
                  $ref: "#/components/schemas/WorkflowApplicability"
      responses:
        "201":
          description: Separately authorized applicability scope appended.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WorkflowVersion"}
        "400": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /api/v1/workflows/{workflowID}/versions/{version}/retire:
    post:
      summary: Retire a published workflow version
      tags: [Workflows]
      operationId: retireWorkflowVersion
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/WorkflowID"}
        - {$ref: "#/components/parameters/WorkflowVersion"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: {type: string, minLength: 1, maxLength: 2000}
      responses:
        "200":
          description: Workflow version retired and excluded from applicability.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WorkflowVersion"}
        "403": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: skawld_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token for native/mobile Authorization Code + PKCE.
  parameters:
...
    WorkflowID:
      in: path
      name: workflowID
      required: true
      schema: {type: string, format: uuid}
    WorkflowVersion:
      in: path
      name: version
      required: true
      schema: {type: integer, minimum: 1}
...
  responses:
    Problem:
      description: RFC 9457-style problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    WorkflowApplicability:
      type: object
      additionalProperties: false
      required: [validation_status]
      properties:
        id: {type: string, format: uuid, readOnly: true}
        site_id: {type: string, format: uuid}
        asset_id: {type: string, format: uuid}
        asset_class: {type: string}
        manufacturer: {type: string}
        model: {type: string}
        process_service: {type: string}
        operating_condition: {type: string}
        validation_status:
          type: string
          enum:
            [VALIDATED, LIKELY_APPLICABLE, NOT_VALIDATED, NOT_APPLICABLE]
        approved_by: {type: string, format: uuid, readOnly: true}
        approved_at: {type: string, format: date-time, readOnly: true}
    WorkflowEvidence:
      type: object
      required: [demonstration_id, event_ids]
      properties:
        demonstration_id: {type: string, format: uuid}
        event_ids:
          type: array
          minItems: 1
          items: {type: string, format: uuid}
    WorkflowStep:
      type: object
      required: [id, kind, evidence]
      properties:
        id: {type: string}
        name: {type: string}
        kind: {type: string, enum: [tool, validation, approval]}
        tool_name: {type: string}
        arguments: {type: object, additionalProperties: true}
        evidence:
          type: array
          items: {$ref: "#/components/schemas/WorkflowEvidence"}
    WorkflowVersion:
      type: object
      required:
        [workflow_id, workflow_key, name, version, status, site_id,
         asset_class, candidate_digest, tool_catalog_digest,
         source_demonstration_ids, steps, analysis, behavioral_changes,
         learning, applicability, prerequisites, required_competencies,
         reviews, evaluations, improvement_candidates, created_at]
      properties:
        workflow_id: {type: string, format: uuid}
        workflow_key: {type: string}
        name: {type: string}
        description: {type: string}
        version: {type: integer, minimum: 1}
        status:
          type: string
          enum:
            [CANDIDATE, APPROVED, REVIEW_REQUIRED, PUBLISHED, RETIRED, REJECTED]
        site_id: {type: string, format: uuid}
        asset_class: {type: string}
        candidate_digest: {type: string, pattern: '^[a-f0-9]{64}$'}
        tool_catalog_digest: {type: string}
        source_demonstration_ids:
          type: array
          minItems: 2
          items: {type: string, format: uuid}
        steps:
          type: array
          minItems: 1
          items: {$ref: "#/components/schemas/WorkflowStep"}
        analysis:
          type: object
          additionalProperties: true
          description: Deterministic sequence variants, conflicts, parameters, and findings.
        behavioral_changes:
          type: object
          additionalProperties: true
          description: Behavior-focused diff from the latest stored version.
        learning:
          type: object
          additionalProperties: true
        applicability:
          type: array
          items: {$ref: "#/components/schemas/WorkflowApplicability"}
        prerequisites:
          type: array
          items: {type: string}
        required_competencies:
          type: array
          items: {type: string}
        effective_at: {type: string, format: date-time}
        review_at: {type: string, format: date-time}
        reviews:
          type: array
          items:
            type: object
            additionalProperties: true
        evaluations:
          type: array
          items:
            type: object
            additionalProperties: true
        improvement_candidates:
          type: array
          description: Human corrections awaiting review; never automatic mutations.
          items:
            type: object
            additionalProperties: true
        created_at: {type: string, format: date-time}
        published_at: {type: string, format: date-time}
        published_by: {type: string, format: uuid}
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string